BLD: Restructure package to separate dependencies
